% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePAM.r
\name{pam_to_raster}
\alias{pam_to_raster}
\title{Make raster object}
\usage{
pam_to_raster(pam, plot_map = TRUE)
}
\arguments{
\item{pam}{a pam object produced by sf_to_pam}

\item{plot_map}{logical. If `TRUE` plot the raster as a map}
}
\value{
A a raster object.
}
\description{
Creates a raster from a PAM
}
\details{
TBC
}
\examples{

}
