% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePAM.r
\name{sf_to_pam}
\alias{sf_to_pam}
\title{Presence-absence matrices}
\usage{
sf_to_pam(
  sf_object,
  taxon_names = NULL,
  resolution = 100,
  sampling = "conservative",
  file_name = NULL,
  crs = "+proj=cea +lat_ts=30 +units=km",
  clip_to_world = "none",
  returnPAM = TRUE
)
}
\arguments{
\item{sf_object}{an object of class "sf" containing all of the polygon data to be converted to a PAM}

\item{taxon_names}{column name in the sf object specifying taxon names to be used to create the PAM}

\item{resolution}{required resolution for the PAM. The appropriate size depends on your data and question. Default is 100 (km) for a Berhman equal area grid (see `crs` below).}

\item{sampling}{one of "conservative" (the default but not necessarily the best), "liberal", or a numeric value from 1:100}

\item{file_name}{path to output file to write the PAM to file as R data (.rds). If no path is specified, the file will not be written.}

\item{crs}{coordinate reference system for the PAM. Default is "+proj=cea +lat_ts=30 +units=km" (a Berhman equal area projection)}

\item{clip_to_world}{One of "none", "sampling", or "liberal". None does no clipping and returns the full PAM. "sampling" and "liberal" sets grid cells with no land to NA.}

\item{returnPAM}{logical. If `TRUE` the PAM is returned and can be used direction.}
}
\value{
A list of two elements. The first element is the PAM and the second is an empty raster object with the same crs and extent as the PAM. This is needed for plotting the PAM as e.g. a species richness map.
}
\description{
Creates a presence-absence matrix (PAM) from an sf object
}
\details{
TBC
}
\examples{

}
